+++ src/components/Catalogues/CatalogueMangaCard.js
