+++ src/components/Catalogues/index.js
@@ -6,9 +6,9 @@ import { Helmet } from "react-helmet";
 import isEmpty from "lodash/isEmpty";
 import { makeStyles } from "@material-ui/styles";
 import Typography from "@material-ui/core/Typography";
-import CatalogueMangaCard from "components/Catalogue/CatalogueMangaCard";
+import CatalogueMangaCard from "components/Catalogues/CatalogueMangaCard";
 import DynamicSourceFilters from "components/Filters/DynamicSourceFilters";
-import CatalogueHeader from "components/Catalogue/CatalogueHeader";
+import CatalogueHeader from "components/Catalogues/CatalogueHeader";
 import CenteredLoading from "components/Loading/CenteredLoading";
 import FullScreenLoading from "components/Loading/FullScreenLoading";
 import { selectIsSourcesLoading, selectSources } from "redux-ducks/sources";
